// src/api-spec/openapi/swagger/data-contracts.ts
export enum V1TradeType {
  TRADE_TYPE_BUY = 'TRADE_TYPE_BUY',
  TRADE_TYPE_SELL = 'TRADE_TYPE_SELL',
}